+++ src/main/kotlin/net/ndrei/teslapoweredthingies/integrations/crafttweaker/PoweredThingiesTweaker.kt
@@ -6,6 +6,7 @@ import net.ndrei.teslapoweredthingies.machines.compoundmaker.CompoundMakerTweake
 import net.ndrei.teslapoweredthingies.machines.fluidburner.FluidBurnerCoolantTweaker
 import net.ndrei.teslapoweredthingies.machines.fluidburner.FluidBurnerFuelTweaker
 import net.ndrei.teslapoweredthingies.machines.fluidcompoundproducer.FluidCompoundProducerTweaker
+import net.ndrei.teslapoweredthingies.machines.incinerator.IncineratorTweaker
 import net.ndrei.teslapoweredthingies.machines.itemcompoundproducer.ItemCompoundProducerTweaker
 import stanhebben.zenscript.annotations.ZenClass
 import stanhebben.zenscript.annotations.ZenMethod
@@ -29,4 +30,7 @@ object PoweredThingiesTweaker {
 
     private val itemCompoundProducerInstance = ItemCompoundProducerTweaker()
     @ZenMethod @JvmStatic fun itemCompoundProducerTweaker() = this.itemCompoundProducerInstance
+
+    private val incineratorInstance = IncineratorTweaker()
+    @ZenMethod @JvmStatic fun incineratorTweaker() = this.incineratorInstance
 }